Update remaining Direct3D 9 samples

+++ Direct3D/ShadowVolume/ShadowVolume.cpp
@@ -373,9 +373,9 @@ HRESULT GenerateShadowMesh( IDirect3DDevice9* pd3dDevice, ID3DXMesh* pMesh, ID3D
                         int nIndex;
                         int nVertIndex[3] =
                         {
-                            pdwPtRep[pdwIBData[f * 3]],
-                            pdwPtRep[pdwIBData[f * 3 + 1]],
-                            pdwPtRep[pdwIBData[f * 3 + 2]]
+                            static_cast<int>(pdwPtRep[pdwIBData[f * 3]]),
+                            static_cast<int>(pdwPtRep[pdwIBData[f * 3 + 1]]),
+                            static_cast<int>(pdwPtRep[pdwIBData[f * 3 + 2]])
                         };
                         nIndex = FindEdgeInMappingTable( nVertIndex[0], nVertIndex[1], pMapping, dwNumEdges );
